+++ src-tauri/src/internal/mod.rs
@@ -1,4 +1,4 @@
 pub mod article; // Declare the articles modules
 pub mod api; // Declare the api module
 pub mod feed_config; // Declare the configuration modules
-pub mod rss; // Declare the rss modules
+pub mod feed; // Declare the rss modules
